requirements.txt
insightface
glob
cv2
Pillow
torch
torchvision

